Added parsing of Nokia Suite .csv file.
Added the most essential parts for parsing .csv file created by Nokia Suite in a way that should be compatible with the rest of the software. To incorporate it fully probably additional argument should be parsed (input_type) and kind of a dropdown list added on web interface. Also small variables changes should be made to keep everything clear.